simplify setting dependencies to simplify invalidating bindings in a model
Dependencies had been "injected" by defining a getter that was returning an instance from the registry. That was causing issues with binding to a "injected" dependency as it was never really set, hence  "changed" event was not being dispatched. Now dependencies are set using setters/getters of the instance, with model changed-event-based-setters it triggers bindings properly now, with no need to invalidate them on inject_init